pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yomi-core"
version = "0.1.0"
description = "The Universal Hybrid-Engine Manga Downloader"
readme = "README.md"
authors = [
    { name = "OmurEKiraz", email = "omrkrz0678@gmail.com" }
]
license = { file = "LICENSE" }
requires-python = ">=3.8"
keywords = ["manga", "downloader", "scraper", "cli", "async", "comics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
]
dependencies = [
    "click>=8.0.0",
    "rich_click>=1.7.0",
    "aiohttp>=3.8.0",
    "requests>=2.28.0",
    "beautifulsoup4>=4.11.0",
    "lxml>=4.9.0",
    "rich>=12.0.0",
    "pillow>=9.0.0",
    "img2pdf>=0.4.0"
]

[project.urls]
Homepage = "https://github.com/OmurEKiraz/yomi-core"
Repository = "https://github.com/OmurEKiraz/yomi-core"
Issues = "https://github.com/OmurEKiraz/yomi-core/issues"

[project.scripts]
yomi = "yomi.cli:cli"